Added postgresql configuration and project structure
