+++ src/main/resources/application-dev.properties
@@ -1,0 +1,10 @@
+app.greeting=IMDB Dev
+
+#JPA related
+spring.datasource.url = jdbc:postgresql://ec2-52-3-200-138.compute-1.amazonaws.com/d3fdeqv2mbgqmi
+spring.jpa.show-sql=true
+spring.datasource.username=qrmiqvglydgzds
+spring.datasource.password=c14e44d26636f618b273c629bbcf05b0a7c091dc83864cad438cd0280d53c176
+spring.sql.init.platform=postgresql
+
+spring.sql.init.mode=always
