+++ src/main/resources/application-prod.properties
@@ -1,0 +1,8 @@
+app.greeting=IMDB
+
+#JPA related
+spring.datasource.url = jdbc:postgresql://ec2-52-3-200-138.compute-1.amazonaws.com/dbmi0q6qjff2i6
+spring.jpa.show-sql=true
+spring.datasource.username=kyvnuukwfkuuuf
+spring.datasource.password=3f4f40d36417c01680524b8d820fa8631bbd0a866ffe640c753cb4387eb0beed
+spring.sql.init.platform=postgresql
