+++ src/main/resources/application.properties
@@ -1,1 +1,6 @@
+spring.profiles.active=dev
+#spring.profiles.active=prod
 
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.generate-ddl=false
+spring.jpa.properties.hibernate.format_sql=true
