+++ src/main/resources/schema-postgresql.sql
@@ -1,0 +1,1 @@
+
