+++ src/test/java/com/group5/IMDB/ImdbApplicationTests.java
@@ -1,7 +1,5 @@
 package com.group5.IMDB;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -15,15 +13,5 @@ class ImdbApplicationTests {
 	@Test
 	void contextLoads() {
 	}
-
-	@Test
-	void welcome() {
-        assertEquals( "Welcome to IMDB", imdbApplication.welcome());
-    }
-
-    @Test
-    void version() {
-        assertEquals( "The current version is 0.0.1", imdbApplication.version());
-    }
 
 }
